use an event handler to log messages rather than passive handler
Previously, we were doing some fudgery to log messages as the result of
commands from the meetings plugin. (or from other maubot plugins)
This was caused by the passive handler not processing messages from the
bot itself, and the passive handler not processing messages of type
NOTICE (which maubot's evt.reply sends messages as)

So this commit removes all the fudgery, and adds a event handler on all
room messages of type NOTICE and TEXT to log messages.

Additionally, this commit moves the tags commands to a programmatic
approach, allowing difference instances to define their own tag / emoji
pairs

Signed-off-by: Ryan Lerch <rlerch@redhat.com>

Add more tags

Signed-off-by: Ryan Lerch <rlerch@redhat.com>